added images and other functionality to posts

+++ data/posts.js
@@ -6,6 +6,13 @@ import { postHelpers, getCurrentDateTime, validators } from "../helper.js";
 export const createPost = async (postObj) => {
   if (postObj?.userID === undefined) throw `Need User ID.`;
   postObj.userID = postHelpers.isuserIDValid(postObj.userID);
+  const user = await users();
+  const existing_user = await user.findOne({
+    _id: new ObjectId(postObj.userID),
+  });
+  if (!existing_user) {
+    throw "User with same user ID doesnot exists";
+  }
   if (postObj?.title === undefined) throw `Title needs to be Provided.`;
   if (postObj?.content === undefined) throw `Content needs to be Provided.`;
   if (postObj?.type === undefined) throw `Type needs to be Provided.`;
@@ -20,12 +27,13 @@ export const createPost = async (postObj) => {
     postObj.lostfoundDetails = postHelpers.islfdetailsValid(
       postObj.lostfoundDetails
     );
+    if (postObj?.image === undefined) throw `Image needs to be Provided.`;
   } else {
     if (postObj.lostfoundDetails)
       throw `General Post cannot have Lost-Found Details.`;
   }
 
-  const updatedtime = getCurrentDateTime();
+  const updatedtime = new Date();
 
   const newPost = {
     title: postObj.title,
@@ -44,13 +52,22 @@ export const createPost = async (postObj) => {
     throw `Could not add Post`;
 
   const newId = insertPost.insertedId.toString();
+  existing_user.posts.push(newId.toString());
+  await user.updateOne(
+    { _id: newPost.userID },
+    { $set: { posts: existing_user.posts } }
+  );
+
   //const post = await get(newId);
   return newId;
 };
 
 export const getAllGeneralPosts = async () => {
   const postsCollection = await posts();
-  let postList = await postsCollection.find({ type: "general" }).toArray();
+  let postList = await postsCollection
+    .find({ type: "general" })
+    .sort({ updatedtime: -1 })
+    .toArray();
   if (!postList) throw `Could not get all Posts`;
   if (postList.length === 0) return postList;
   postList = postList.map((element) => {
@@ -64,6 +81,7 @@ export const getAllLFPosts = async () => {
   const postsCollection = await posts();
   let postList = await postsCollection
     .find({ type: { $in: ["lost", "found"] } })
+    .sort({ updatedtime: -1 })
     .toArray();
   if (!postList) throw `Could not get all Posts`;
   if (postList.length === 0) return postList;
@@ -88,6 +106,7 @@ export const getmyPosts = async (id) => {
   const postsCollection = await posts();
   let postList = await postsCollection
     .find({ userID: new ObjectId(id) })
+    .sort({ updatedtime: -1 })
     .toArray();
   if (!postList) throw `Could not get My Posts`;
   if (postList.length === 0) return postList;
@@ -98,8 +117,16 @@ export const getmyPosts = async (id) => {
   return postList;
 };
 
-export const removePost = async (id) => {
+export const removePost = async (id, userID) => {
   validators.checkId(id);
+  validators.checkId(userID);
+  const user = await users();
+  let existing_user = await user.findOne({
+    _id: new ObjectId(userID),
+  });
+  if (!existing_user) {
+    throw "User with same user ID doesnot exists";
+  }
   const postsCollection = await posts();
   const deletionInfo = await postsCollection.findOneAndDelete({
     _id: new ObjectId(id),
@@ -107,6 +134,7 @@ export const removePost = async (id) => {
   if (!deletionInfo) {
     throw `Could not delete post with id of ${id}`;
   }
+  await user.updateOne({ _id: new ObjectId(userID) }, { $pull: { posts: id } });
   return deletionInfo;
 };
 
@@ -121,29 +149,3 @@ export const removePostbyUser = async (id) => {
   }
   return deletionInfo.deletedCount;
 };
-
-// try {
-//   const newPost = {
-//     title: "New Post Title",
-//     content: "New Post Content",
-//     image: "Image",
-//     type: "lost",
-//     lostfoundDetails: {
-//       location: "Hoboken",
-//       date: "02/02/2024",
-//       contact_info: "84957485349",
-//       pet_details: {
-//         species: "dog",
-//         breed: "whatever",
-//         color: "red",
-//         other_details: "it is a cat",
-//       },
-//     },
-//     userID: "65d69f9de5cf3db66f90a055",
-//   };
-
-//   let id = await getPostsbyID("661b3b2ef8a3a24c68bbde82");
-//   console.log(id);
-// } catch (e) {
-//   console.log(e);
-// }
